+++ IM15_Firesphere_199/Managers/InputManager.h
@@ -26,7 +26,7 @@ class InputManager
 		static void zoomIn();
 		static void zoomOut();
 		static void Rotate();
-		static void Rotate(double dx, double dy);
+		static void Rotate(float dx, float dy);
 
 		static void UpCamera();
 		static void DownCamera();
